refactor(interfaces): use kebab-case to name file
