+++ projects/ng-snotify/src/lib/enums/snotify-position.enum.ts
