+++ projects/ng-snotify/src/lib/enums/snotify-style.enum.ts
