+++ projects/ng-snotify/src/lib/interfaces/index.ts
@@ -1,8 +1,8 @@
-export * from './SnotifGlobalConfig.interface';
-export * from './Snotify.interface';
-export * from './SnotifyAnimate.interface';
-export * from './SnotifyButton.interface';
-export * from './SnotifyDefaults.interface';
-export * from './SnotifyNotifications.interface';
-export * from './SnotifyStyles.interface';
-export * from './SnotifyToastConfig.interface';
+export * from './snotif-global-config.interface';
+export * from './snotify.interface';
+export * from './snotify-animate.interface';
+export * from './snotify-button.interface';
+export * from './snotify-defaults.interface';
+export * from './snotify-notifications.interface';
+export * from './snotify-styles.interface';
+export * from './snotify-toast-config.interface';
