+++ projects/ng-snotify/src/lib/interfaces/snotif-global-config.interface.ts
