+++ projects/ng-snotify/src/lib/interfaces/snotify-animate.interface.ts
