+++ projects/ng-snotify/src/lib/interfaces/snotify-button.interface.ts
