+++ projects/ng-snotify/src/lib/interfaces/snotify-notifications.interface.ts
