+++ projects/ng-snotify/src/lib/interfaces/snotify-styles.interface.ts
