+++ projects/ng-snotify/src/lib/models/snotify-toast.model.ts
@@ -1,7 +1,7 @@
-import { SnotifyToastConfig } from '../interfaces/SnotifyToastConfig.interface';
+import { SnotifyToastConfig } from '../interfaces/snotify-toast-config.interface';
 import { Subject, Subscription } from 'rxjs';
 import { SnotifyEvent } from '../types/event.type';
-import { SnotifyStyle } from '../enums/SnotifyStyle.enum';
+import { SnotifyStyle } from '../enums/snotify-style.enum';
 // @TODO remove method in observable way
 /**
  * Toast main model
