+++ projects/ng-snotify/src/public-api.ts
@@ -9,5 +9,5 @@ export * from './lib/models';
 export * from './lib/pipes';
 export * from './lib/services';
 export * from './lib/snotify.module';
-export * from './lib/toastDefaults';
+export * from './lib/toast-defaults';
 export * from './lib/types';
